rename support files
