+++ cypress-tests/cypress/e2e/v2/anonymousNavigation.cy.ts
@@ -3,13 +3,13 @@ import {
   getUserData,
   validateLoginV2,
 } from "../../support/commands/general";
-import { User } from "../../support/types/user.types";
+import { User } from "../../support/types/user";
 import {
   createProjectIfMissingAPIV2,
   deleteProjectFromAPIV2,
   getProjectByNamespaceAPIV2,
-} from "../../support/utils/projectsV2.utils";
-import { verifySearchIndexing } from "../../support/utils/search.utils";
+} from "../../support/utils/projects";
+import { verifySearchIndexing } from "../../support/utils/search";
 
 const anonymousSession = {
   id: ["anonymousNavigation-anonymousUser", getRandomString()],
