+++ cypress-tests/cypress/e2e/v2/dataConnectors.cy.ts
@@ -3,17 +3,17 @@ import {
   getUserData,
   validateLoginV2,
 } from "../../support/commands/general";
-import { ProjectIdentifierV2 } from "../../support/types/project.types";
-import { User } from "../../support/types/user.types";
+import { ProjectIdentifierV2 } from "../../support/types/projects";
+import { User } from "../../support/types/user";
 import {
   createProjectIfMissingAPIV2,
   deleteProjectFromAPIV2,
   getProjectByNamespaceAPIV2,
-} from "../../support/utils/projectsV2.utils";
+} from "../../support/utils/projects";
 import {
   createGroupIfMissingAPI,
   deleteGroupFromAPI,
-} from "../../support/utils/group.utils";
+} from "../../support/utils/groups";
 
 const sessionId = ["dataConnectors", getRandomString()];
