+++ cypress-tests/cypress/support/types/user.ts
