+++ cypress-tests/cypress/support/utils/search.ts
